+++ src/components/common/quizAnswer/QuizAnswer.Style.ts
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 import Colors from '../../../styles/Colors';
-import FontStyles from '../../../styles/FontStyle';
+import FontStyles from '../../../styles/FontStyles';
 
 export const QuizAnswer = styled.div<{ $color: 'gray' | 'purple' }>`
   ${FontStyles.SM_Medium}
