+++ src/components/home/cartoon/Cartoon.Style.ts
@@ -1,7 +1,7 @@
 import styled from 'styled-components';
 import RightArrowIcon from '../../../assets/icons/RightArrowIcon.svg?react';
 import Colors from '../../../styles/Colors';
-import FontStyles from '../../../styles/FontStyle';
+import FontStyles from '../../../styles/FontStyles';
 
 export const Cartoon = styled.div`
   display: flex;
