+++ src/components/newsDetail/aiSummary/AISummary.Style.ts
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 import Colors from '../../../styles/Colors';
-import FontStyles from '../../../styles/FontStyle';
+import FontStyles from '../../../styles/FontStyles';
 
 export const AIPart = styled.div`
   width: 44.5625rem;
