+++ src/components/register/errorMessage/ErrorMessage.Sytle.ts
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 import Colors from '../../../styles/Colors';
-import FontStyles from '../../../styles/FontStyle';
+import FontStyles from '../../../styles/FontStyles';
 
 export const ErrorMessage = styled.div`
   display: flex;
