+++ src/components/register/selectItem/SelectItem.Style.ts
@@ -1,7 +1,7 @@
 import styled from 'styled-components';
 import XIcon from '../../../assets/icons/XIcon.svg?react';
 import Colors from '../../../styles/Colors';
-import FontStyles from '../../../styles/FontStyle';
+import FontStyles from '../../../styles/FontStyles';
 
 export const Keyword = styled.div`
   ${FontStyles.XS_Medium}
